requirements.txt
torch>=2.3.0
torchvision>=0.18.0
gymnasium>=0.29.1
minigrid>=2.4.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
tqdm>=4.66.0
PyYAML>=6.0.1
Pillow>=10.0.0
huggingface-hub>=0.23.0
transformers>=4.41.0
safetensors>=0.4.3
